update rotate for Collection
